Remove double-meaning of EBUSY in queue_destroy
If the mutex cannot be destroyed, then that error is no longer reported
from task_queue_destroy().  This is because pthread_mutex_destroy()
could return EBUSY, which means that if the mutex cannot be destroyed,
we would be returning EBUSY but we would have left the queue in an
undefined state, which our contract forbids.